+++ Scenes and Scripts/Scenes/Player/Inventory/InventoryDropable_SCRIPT.gd
@@ -85,7 +85,9 @@ func _ready():
 		ITEM_TYPE_Label.text = "Blank Flower"
 	elif "PINKFLOWER" in ITEM_TYPE:
 		ITEM_TYPE_Label.text = "Pink Flower"
-	
+	elif "YELLOWFLOWER" in ITEM_TYPE:
+		ITEM_TYPE_Label.text = "Yellow Flower"
+
 	elif "EFFICIENCYPOTION" in ITEM_TYPE:
 		ITEM_TYPE_Label.text = "Efficiency Potion"
 	elif "HASTEPOTION" in ITEM_TYPE: