+++ Scenes and Scripts/Scenes/The Island/Explorer Note/ExplorerNote_StaticBody3D_SCRIPT.gd
@@ -50,6 +50,7 @@ extends StaticBody3D
 @export var root : Node3D
 
 func on_raycast_hit_explorer_note():
+	ExplorerNotesManager.CurrentlyInteracting_ID = root.NoteID
 	InteractionManager.is_hovering_over_explorer_note = true
 
 func on_raycast_exit_explorer_note():