+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 yellow flower text inside inventory is now "Yellow Flower" instead of "Yellowflower"
 fixed a bug where you could close the workbench by pressing "E", then being able to open it again by pressing "E"
 fixed 3d inventory pickup items noclipping through the ground
+there are now 16 slots in inventory pockets instead of 9 to store items
 
 
 ---