+++ CHANGELOG.md
@@ -8,6 +8,7 @@ yellow flower text inside inventory is now "Yellow Flower" instead of "Yellowflo
 fixed a bug where you could close the workbench by pressing "E", then being able to open it again by pressing "E"
 fixed 3d inventory pickup items noclipping through the ground
 there are now 16 slots in inventory pockets instead of 9 to store items
+change player collision shape to reduce lag
 
 
 ---