+++ CHANGELOG.md
@@ -11,7 +11,9 @@ there are now 16 slots in inventory pockets instead of 9 to store items
 there are now 49 slots in the chest instead of 25 to store items
 an all new crafting system which can be accessed by interacting with the workbench in the testing area
 change player collision shape to reduce lag
-new stone pickup item inventory
+new stone inventory item
+new wood plank inventory item
+
 
 
 ---