+++ CHANGELOG.md
@@ -15,6 +15,7 @@ new stone inventory item
 new wood plank inventory item
 an all new explorer notes system, with the ability to collect notes and view them in your inventory
 removed the old hand item system
+changed the FOV cap from 60-120 to 90-120
 [added a hotbar to the player, where you can quickly access and hold items]
 [added the ability to build]